src/backend/requirements.txt
aiohttp==3.9.3
python-dotenv==1.0.1
ujson==5.9.0
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
SQLAlchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
clickhouse-driver==0.2.6
kafka-python==2.0.2
prometheus-client==0.20.0
structlog==24.1.0 